requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
openai
httpx
tqdm
prometheus-client
asyncio
python-dotenv
pytest
pytest-asyncio
pathlib
watchdog
apscheduler
# ast is part of Python's standard library and is intentionally not installed from PyPI.
